+++ test/unit/json_ld_test.exs
@@ -1,0 +1,5 @@
+defmodule JSON.LDTest do
+  use ExUnit.Case
+
+  doctest JSON.LD
+end